feat(ia): batch 6 — team page restructure
Per Batch 6 of admin-ia-redesign-plan.md.

- Decompose 380-line /members/add/page.tsx into 3 reusable components
  in src/components/team/members/ (round-1 F10/T8):
  - team-add-from-tenant-section.tsx (search + role select + add)
  - team-invite-by-email-section.tsx (email + role + invite link)
  - team-pending-invitations-list.tsx (pending invites with cancel + copy)
  Plus new __tests__/ for each (3 tests × 2 behaviors = 6 tests; existing
  add page had no unit-test coverage — first-time coverage added).
- Inline /members/list/page.tsx content into /members/page.tsx; the
  Members page now serves the roster directly with:
  - "メンバーを追加" button → Dialog hosting AddFromTenant + InviteByEmail
  - "オーナー権限を移譲" link → /members/transfer-ownership
  - PendingInvitationsList rendered below the roster
- git rm /members/{list,add}/.
- git mv /members/transfer/ → /members/transfer-ownership/ (renamed for
  clarity per plan's destructive-action UX guidance).
- git mv 3 security pages to top-level: /policy, /key-rotation, /webhooks.
- Add SectionLayout wrappers for each (round-1 F3): policy/layout.tsx,
  key-rotation/layout.tsx, webhooks/layout.tsx.
- git rm now-orphaned /security/{layout.tsx, page.tsx}.
- TeamRotateKeyButton: add typed "rotate" confirmation input gate before
  the destructive action (round-1 S4 — required because the new IA
  reduces clicks from 3 → 1).

Tests: 74 passed (admin sidebar + admin scope + team members + team
policy settings + team webhook card + team-rotate-key button bulk).
