+++ src/app/[locale]/admin/teams/[teamId]/key-rotation/layout.tsx
@@ -1,17 +1,17 @@
 "use client";
 
 import { useTranslations } from "next-intl";
-import { Shield } from "lucide-react";
+import { KeyRound } from "lucide-react";
 import { SectionLayout } from "@/components/settings/account/section-layout";
 
-export default function TeamSecurityLayout({ children }: { children: React.ReactNode }) {
+export default function TeamKeyRotationLayout({ children }: { children: React.ReactNode }) {
   const t = useTranslations("AdminConsole");
 
   return (
     <SectionLayout
-      icon={Shield}
-      title={t("teamSectionSecurity")}
-      description={t("teamSectionSecurityDesc")}
+      icon={KeyRound}
+      title={t("teamSectionKeyRotation")}
+      description={t("teamSectionKeyRotationDesc")}
     >
       {children}
     </SectionLayout>
