+++ src/app/[locale]/admin/teams/[teamId]/key-rotation/page.tsx
